pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "offensive-langgraph-playground"
version = "0.0.1"
dependencies = [
    "fabric>=3.2.2",
    "invoke>=2.2.0",
    "langgraph>=0.2.35",
    "langchain-openai>=0.2.2",
    "mako>=1.3.5",
    "python-dotenv>=1.0.1"
]
requires-python = ">= 3.8"